<No Comment Entered>

+++ src/com/dexels/navajo/document/nanoimpl/NavajoFactoryImpl.java
@@ -75,7 +75,7 @@ public class NavajoFactoryImpl extends NavajoFactory {
   }
   public Header createHeader(Navajo n, String rpcName, String rpcUser, String rpcPassword, long expiration_interval) {
     HeaderImpl hi = new HeaderImpl(n,rpcUser,rpcPassword,rpcName);
-    hi.setExpiration((int)expiration_interval);
+    hi.setExpiration(expiration_interval);
     return hi;
 //    throw new java.lang.UnsupportedOperationException("Method createHeader() not yet implemented.");
   }
